+++ recipe/symfony.php
@@ -106,7 +106,7 @@ task('deploy:assetic:dump', function () {
  * Clear Cache
  */
 task('deploy:cache:clear', function () {
-    run('{{env_vars}} {{bin/php}} {{bin/console}} cache:clear {{console_options}} --no-debug --no-warmup');
+    run('{{env_vars}} {{bin/php}} {{bin/console}} cache:clear {{console_options}} --no-warmup');
 })->desc('Clear cache');
 
 /**