+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## master
 [v4.3.0...master](https://github.com/deployphp/deployer/compare/v4.3.0...master)
 
---
+### Added
+- Added possibility to use callable when setting 'default_stage'.
 
 ## v4.3.0
 [v4.2.1...v4.3.0](https://github.com/deployphp/deployer/compare/v4.2.1...v4.3.0)