+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # Changelog
 
-## master
-[v4.3.0...master](https://github.com/deployphp/deployer/compare/v4.3.0...master)
+## 4.x
+[v4.3.1...4.x](https://github.com/deployphp/deployer/compare/v4.3.1...4.x)
+
+-
+
+## v4.3.1
+[v4.3.0...v4.3.1](https://github.com/deployphp/deployer/compare/v4.3.0...v4.3.1)
 
 ### Added
 - Added possibility to use callable when setting 'default_stage'.
+
 
 ## v4.3.0
 [v4.2.1...v4.3.0](https://github.com/deployphp/deployer/compare/v4.2.1...v4.3.0)